MWPW-141920 refactor branch links
